requirements.txt
# python-telegram-bot==22.0
# requests>=2.28.1
# python-dotenv>=1.0.0
# motor>=3.1.1,<4
# pymongo>=4.1,<5
python-telegram-bot==22.0
requests>=2.28.1
python-dotenv>=1.0.0
firebase-admin>=6.0.0
apscheduler>=3.10.1
